Changed Scenario so that it can be parsed from JSON instead of XML.

+++ src/main/java/de/uni_potsdam/hpi/bpt/bp2014/jcomparser/xml/DomainModel.java
@@ -87,7 +87,7 @@ public class DomainModel implements IDeserialisableJson, IPersistable {
 		try {
 			this.domainModelJson = new JSONObject(element);
 
-			this.domainModelModelID = this.domainModelJson.getInt("_id");
+			this.domainModelModelID = this.domainModelJson.getLong("_id");
 			this.versionNumber = this.domainModelJson.getInt("revision");
 
 			generateDataClasses();
